Fix quality stats on gems not showing up in gem tooltip (#1302)
* Fix quality stats on gems not showing up in gem tooltip

We were still using the old logic for secondaryGrantedEffect that PoB 1 uses, so if any skill had a quality stat on a additionalGrantedEffects, it wouldn't show up in the quality tooltip or get added to the main gem tooltip
We now iterate over the first 2 entries in the additionalGrantedEffects list to check for quality stats in them
Note: We technically handle quality stats wrong for quite a few gem. Some gems have restrictions on the stat sets the quality stat should apply to but we currently just apply the quality stat to all the stat sets.
I'll look into fixing this in the future
We were also missing the export of the quality stat on Hand of Chayula

* Fix gem display

* Fix tooltip display

---------

Co-authored-by: LocalIdentity <localidentity2@gmail.com>