Add support for "Walker of the Wilds" Keystone and other "leech recovers based on x" mods (#1348)
* Refactor leech calculation

Intent was to standardize mod phrasing and logic, as well as reducing
repetition of code.

* Add `hasTag` function to `modLib` in `ModTools.lua`

There was no easy existing way to check if a specific tag is already
present on a on mod (Or I couldn't find it)
So I added a function for it, which makes it easy to check for, add, or
replace/modify specific tags

* Add option to override condition values via `cfg`

It was kinda hard to tell `ModStore:GetCondition` to assume a condition
is `false` without changing the actual conditions on the actor, so I
added the opton to force the value via `cfg.overrideCond`.

I needed to force the condition because I wanted to tabulate mods that
do not apply, but are used as basis for the value of other mods. e.g.
Leech Elemental Damage *instead* of Physical Damage

* Improve leech conversion refactor

Better handling of actual conversion of mods with the correct tags for
display in breakdown later

* Add new leech modifiers to relevant breakdowns

* Add parsing for "Walker of the Wilds" Keystone

* Add support for further leech conversion mods

This adds general support for parsing most expected forms of leech
conversion mods.

Specific mods now supported:
- Sap of Nightmares (Acolyte of Chayula)
- Mystic Harvest (Amazon)
- Spire of Ire (Unique Helix Spear)

* Add support for Energy Shield leech mods

Mostly just for testing, as these mods don't exist in the game yet, but
might be added in the future via unique items.